docs: :zap: Changing array messages to number strings
Number strings can be translated directly to the desired arrays in binary. 8 = 1000. Changed the message method back to the UDP method from tutorial sheets and removed start command as Simulink will do this

Initially not working due to target 1 not being in frame to start. Need to update script to uphold this